Support for new shortcode parameters
Added show_filters parameter to sermons shortcode.
Verify and change boolean shortcode parameters to true/false.
Changed filter selection to $_POST instead of $_GET.
Added support for $_POST filter selection.
Changed filter selection to show in the same page instead of pulling a taxonomy view so that any shortcode parameters are retained.
Added support in content-sermon-single to not display the pagination if the single is being rendered as part of the show_initial in the archive views.